% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_biodiversidad.R
\name{BD_biodiversidad}
\alias{BD_biodiversidad}
\title{Apéndice BD biodiversidad}
\usage{
BD_biodiversidad(
  BD_flora,
  sp,
  nom_ssubc,
  portada = "default",
  portada_opts = NULL
)
}
\arguments{
\item{BD_flora}{dataframe con la base de datos de las parcelas de inventarios florísticos.}

\item{sp}{especie en categoría de conservación objeto del informe.}

\item{nom_ssubc}{Nombre de la subsubcuenca}

\item{portada}{portada para los apendices. Disponible: `default`, `MLP612`, `KIM753` u `otra`.}

\item{portada_opts}{opciones para personalizar una portada. ver `details`.}
}
\value{
lista con resultados asociados al apendice de biodiversidad
}
\description{
Apéndice BD biodiversidad
}
\details{
En el parametro `portada_opts` debe ingresar una lista con los argumentos de la funcion `portada_opts()`.
}
