% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_inventarios.R
\name{BD_inventarios}
\alias{BD_inventarios}
\title{Apéndice BD inventarios forestales}
\usage{
BD_inventarios(
  BD_fore,
  BNP_cuenca,
  sp,
  portada = "default",
  portada_opts = NULL
)
}
\arguments{
\item{BD_fore}{dataframe con la base de datos de las parcelas de inventarios forestales.}

\item{BNP_cuenca}{capa vectorial con los BNP en la cuenca. Obtener con `get_BNP_cuenca()`.}

\item{sp}{especie en categoría de conservación objeto del informe.}

\item{portada}{portada para los apendices. Disponible: `default`, `MLP612`, `KIM753` u `otra`.}

\item{portada_opts}{opciones para personalizar una portada. ver `details`.}
}
\value{
lista con resultados asociados al apendices de inventarios forestales
}
\description{
Apéndice BD inventarios forestales
}
\details{
En el parametro `portada_opts` debe ingresar una lista con los argumentos de la funcion `portada_opts()`.
}
