% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carto_digital.R
\name{carto_digital}
\alias{carto_digital}
\alias{get_cuenca}
\alias{get_caminos}
\alias{get_CN}
\alias{get_hidro}
\alias{get_ubicacion}
\alias{get_obras}
\alias{get_BNP_cuenca}
\alias{get_BNP_intervencion}
\alias{get_BNP_alterar}
\alias{get_ECC_alt}
\alias{get_ECC_int}
\alias{get_uso}
\alias{get_veg}
\alias{get_inv_flora}
\alias{get_inv_fores}
\alias{get_prospeccion}
\alias{get_BNP_alt_sin_censo}
\alias{get_BNP_int_sin_censo}
\alias{get_BNP_antes}
\alias{get_BNP_despues}
\alias{get_carto_digital}
\title{Title}
\usage{
get_cuenca(uso_veg)

get_caminos(cuenca)

get_CN(cuenca, dem = NULL)

get_hidro(cuenca, fuente_hidro = c("MOP", "BCN"))

get_ubicacion(obras, cuenca)

get_obras(obras, cuenca)

get_BNP_cuenca(uso_veg, sp)

get_BNP_intervencion(BNP_cuenca, obras, BNP_inter = NULL)

get_BNP_alterar(BNP_alter = NULL, BNP_cuenca, alt_ok = F)

get_ECC_alt(BNP_alter, censo, sp)

get_ECC_int(censo, sp, BNP_inter, BNP_alter = NULL, upto5m = T)

get_uso(uso_veg)

get_veg(uso_veg)

get_inv_flora(BD_flora, BNP_cuenca, bd_lista, in_bnp_obra = T, obras = NULL)

get_inv_fores(BD_fore, BNP_cuenca, bd_lista = TRUE)

get_prospeccion(BD_flora, BD_fore, censo, sp)

get_BNP_alt_sin_censo(BNP_alter, densidad)

get_BNP_int_sin_censo(BNP_inter, densidad)

get_BNP_antes(BNP_cuenca)

get_BNP_despues(BNP_cuenca, BNP_inter, BNP_alter)

get_carto_digital(
  uso_veg,
  sp,
  obras,
  censo,
  BNP_inter = NULL,
  BNP_alter = NULL,
  alt_ok = F,
  densidad = NULL,
  BD_flora = NULL,
  BD_fore = NULL,
  upto5m = T,
  bd_flora_lista = F,
  bd_fore_lista = F,
  in_bnp_obra = T,
  add_cam = F,
  add_hidro = F,
  fuente_hidro = c("MOP", "BCN"),
  add_CN = F
)
}
\arguments{
\item{uso_veg}{capa vectorial de la cartografía de uso y vegetación de la cuenca.}

\item{cuenca}{capa vectorial de la cuenca. obtener con `get_cuenca()`.}

\item{dem}{modelo de elevación digital. Ingresar ruta del archivo tif u objeto stars. Si es \code{NULL}, el dem se obtendra con `elevatr::get_elev_raster()`.}

\item{fuente_hidro}{fuente de la capa hidrografica. `MOP` o `BCN`.}

\item{obras}{capa vectorial con las obras del proyecto.}

\item{sp}{especie en categoría de conservación objeto del informe.}

\item{BNP_cuenca}{capa vectorial con los BNP en la cuenca. Obtener con `get_BNP_cuenca()`.}

\item{BNP_inter}{capa vectorial con la superficie de intervención de BNP. default \code{NULL}.}

\item{BNP_alter}{capa vectorial con la superficie de alteración de BNP. default \code{NULL}.}

\item{alt_ok}{valor logico. \code{TRUE} para indicar que el BNP_alter ingresado esta ok. \code{FALSE} si solo corresponden a las obras que alteran.}

\item{censo}{capa vectorial de tipo POINT con el censo de la especie objetivo.}

\item{upto5m}{valor logico. \code{TRUE} para considerar de individuos con un radio de hasta 5 metros de las obras.}

\item{BD_flora}{dataframe con la base de datos de las parcelas de inventarios florísticos.}

\item{bd_lista, bd_fore_lista, bd_flora_lista}{valor logico. \code{TRUE} si los datos ingresados estan listos.}

\item{in_bnp_obra}{valor logico. \code{TRUE} para filtrar los puntos dentro de BNP con obras.}

\item{BD_fore}{dataframe con la base de datos de las parcelas de inventarios forestales.}

\item{densidad}{número de individuos por hectárea de la especie objetivo (sp).}

\item{add_cam, add_hidro, add_CN}{valores logicos para indicar si agregar caminos, hidrografia y/o curvas de nivel respectivamente}
}
\value{
capas vectoriales de la cartografía digital
}
\description{
Title
}
