% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bd.R
\name{check_bd_flora}
\alias{check_bd_flora}
\alias{check_bd_fore}
\title{Chequeo Bases de Datos}
\usage{
check_bd_flora(x)

check_bd_fore(x)
}
\arguments{
\item{x}{dataframe con la base de datos. Generarla con `prepare_bd_fore_150()`}
}
\value{
Devuelve alertas impresas en la consola o bien en modulos de shiny.
}
\description{
Chequeo Bases de Datos
}
