% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estadisticos.R
\name{estadisticos_PAS150}
\alias{estadisticos_PAS150}
\title{estadisticos_PAS150}
\usage{
estadisticos_PAS150(a, A, data, sp)
}
\arguments{
\item{a}{superficie parcela en hectareas.}

\item{A}{superficie unidad a inventariar. ej: BNP de la cuenca.}

\item{data}{data.frame con los datos del NHA. Obter con `BD_inventarios()`.}

\item{sp}{especie en categoría de conservación objeto del informe.}
}
\value{
data.frame con estadisticos
}
\description{
estadisticos_PAS150
}
