% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sup_ha.R
\name{format_sup_ha}
\alias{format_sup_ha}
\title{Formatear precisión de la superficie en un Shapefile}
\usage{
format_sup_ha(shp, field_sup = "Sup_ha", precision = 2L)
}
\arguments{
\item{shp}{Carácter. Ruta completa al archivo shapefile (.shp).}

\item{field_sup}{Carácter. Nombre del campo que contiene la superficie. Por defecto "Sup_ha".}

\item{precision}{Entero. Número de decimales deseados para el redondeo y metadatos. Por defecto 2.}
}
\value{
Invisibilidad de un valor lógico (TRUE si el proceso fue exitoso, FALSE en caso contrario).
}
\description{
Esta función calcula el área en hectáreas para una capa de polígonos, actualiza
el campo especificado (por defecto "Sup_ha") y modifica binariamente los metadatos
del archivo .dbf asociado para asegurar que la precisión decimal se mantenga
y sea visualizada correctamente en herramientas externas como Google Earth, QGIS o ArcGIS.
}
