% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragStats.R
\name{FragStats}
\alias{FragStats}
\alias{FragStats_pre}
\alias{FragStats_post}
\title{FragStats}
\usage{
FragStats_pre(dir, path_antes, path_despues)

FragStats_post(dir, path_antes, path_despues)
}
\arguments{
\item{dir}{directorio}

\item{path_antes}{ruta del archivo '.shp' del BNP antes del proyecto}

\item{path_despues}{ruta del archivo '.shp' del BNP despues del proyecto}
}
\value{
crea carpetas y archivos del apendice de fragmentación
}
\description{
FragStats
}
