% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alteracion.R
\name{get_alter_bnp_remanente}
\alias{get_alter_bnp_remanente}
\title{Identifica y calcula la superficie de alteración en fragmentos remanentes de BNP}
\usage{
get_alter_bnp_remanente(BNP, obras, sp, sup_ha_max = 0.5)
}
\arguments{
\item{BNP}{Objeto `sf` (polígonos) que representa el Bosque Nativo de Preservación.
Debe contener al menos las columnas `Nom_ssubc` y `BNP_ECC`.}

\item{obras}{Objeto `sf` (polígonos) que representa las áreas de intervención de las obras.}

\item{sp}{String con el nombre científico de la especie objeto de protección (ECC)
para filtrar en la columna `BNP_ECC`.}

\item{sup_ha_max}{Valor numérico que define el umbral máximo de superficie (en hectáreas)
para que un fragmento seccionado sea considerado "alterado". Por defecto es 0.5 ha.}
}
\value{
Un objeto `sf` con los polígonos de BNP remanente que cumplen con el criterio
  de tamaño máximo, etiquetados con la columna `Afectacion`. Si no se encuentran
  remanentes que cumplan los criterios, devuelve `NULL` de forma invisible.
}
\description{
Esta función procesa la capa de Bosque Nativo de Preservación (BNP) para identificar
fragmentos que, tras ser seccionados por las obras del proyecto (intervención),
resultan en superficies remanentes menores a un umbral específico. Estos fragmentos
se clasifican como "Alteración BNP remanente" debido a que su pequeño tamaño
compromete su viabilidad biológica o conectividad.
}
