% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{get_densidad}
\alias{get_densidad}
\title{Obtener densidad media de una especie}
\usage{
get_densidad(BD_fore, sp)
}
\arguments{
\item{BD_fore}{Dataframe con datos de inventario forestal.}

\item{sp}{Nombre de la especie objetivo.}
}
\value{
Densidad media (N/ha) redondeada.
}
\description{
Obtener densidad media de una especie
}
