% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BD_fragmentacion.R
\name{get_lsm_analisis}
\alias{get_lsm_analisis}
\title{Title}
\usage{
get_lsm_analisis(
  sf_uso,
  sf_obras,
  path_frag,
  ECC,
  subusos_noveg,
  alteracion = T,
  spp_acomp = NULL,
  prop = NULL,
  estadisticos = NULL,
  portada = "default",
  portada_opts = NULL
)
}
\arguments{
\item{sf_uso}{objeto sf con uso de suelo}

\item{sf_obras}{objeto sf con las obras}

\item{path_frag}{ruta del archivo `xls` o `xlsx` con los resultados del FragStats}

\item{ECC}{especie objeto del informe}

\item{subusos_noveg}{subusos que no corresponden a vegetación}

\item{alteracion}{valor lógico. \code{TRUE} si existe alteración de hábitat}

\item{spp_acomp}{data.frame con especies acompañantes. obtener con `BD_biodiversidad()`}

\item{prop}{data.frame con las proporciones por estado de desarrollo de la especie objetivo. obtener con `BD_inventarios()`}

\item{estadisticos}{data.frame con los estadísticos de muestreo. obtener con `estadisticos_PAS150()`}

\item{portada}{portada para los apendices. Disponible: `default`, `MLP612`, `KIM753` u `otra`.}

\item{portada_opts}{opciones para personalizar una portada. ver `details`.}
}
\value{
lista con resultados asociados al apéndice de fragmentacion
}
\description{
Title
}
