% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portadas.R
\name{portada}
\alias{portada}
\alias{wb_portada_PAS150}
\alias{portada_opts}
\title{Portadas de apendices}
\usage{
wb_portada_PAS150(
  wb,
  apendice = c("inventarios", "biodiversidad", "fragmentacion"),
  nom_ssubc = NULL,
  opts = portada_opts()
)

portada_opts(
  nom_proj = NULL,
  etapa_proj = "EIA",
  logo = NULL,
  plantilla = "default"
)
}
\arguments{
\item{wb}{Workbook.}

\item{apendice}{apendices correspondiente. `inventarios`, `biodiversidad` o `fragmentacion`.}

\item{nom_ssubc}{nombre de la subsubcuenca.}

\item{opts}{Opciones para personalizar una portada. usar [portada_opts()].}

\item{nom_proj}{Carácter. Nombre oficial del proyecto que se mostrará en la portada.}

\item{etapa_proj}{Carácter. Etapa del proceso de evaluación del proyecto. Las opciones permitidas son \code{"EIA"}, \code{"Ad"}, \code{"AdComp"} y \code{"RF"}. Por defecto es \code{"EIA"}.}

\item{logo}{Carácter. Ruta al archivo de imagen (ej. PNG, JPG, SVG) que contiene el logo institucional del cliente.}

\item{plantilla}{Carácter. Selección de una configuración predefinida para proyectos específicos. 
Las opciones disponibles son \code{"default"}, \code{"KIM753"} (Kimal-Lo Aguirre) y \code{"MLP612"} (Minera Los Pelambres).}
}
\value{
Workbook con portada

Una \code{lista} que contiene los parámetros configurados para ser utilizados por \code{wb_portada_PAS150()}.
}
\description{
Genera una lista de configuración para definir títulos, logos y el tipo de estudio
que se mostrarán en la portada de los libros de Excel generados para los apéndices.
}
\details{
Al seleccionar una \code{plantilla} específica distinta de \code{"default"}, los valores de \code{nom_proj} y \code{logo} 
se ajustarán automáticamente a los parámetros predefinidos del proyecto, ignorando los valores proporcionados manualmente.
}
