% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bd.R
\name{prepare_bd_flora_150}
\alias{prepare_bd_flora_150}
\alias{prepare_bd_fore_150}
\title{Preparar datos}
\usage{
prepare_bd_flora_150(
  BD,
  bd_lista = TRUE,
  BNP_cuenca = NULL,
  in_bnp_obra = FALSE,
  obras = NULL
)

prepare_bd_fore_150(BD, bd_lista = FALSE, BNP_cuenca = NULL)
}
\arguments{
\item{BD}{dataframe o ruta con la base de datos de las parcelas de inventarios.}

\item{bd_lista}{valor logico. \code{TRUE} si los datos ingresados estan listos.}

\item{BNP_cuenca}{capa vectorial con los BNP en la cuenca. Obtener con `get_BNP_cuenca()`.}

\item{in_bnp_obra}{valor logico. \code{TRUE} para filtrar los puntos dentro de BNP con obras.}

\item{obras}{capa vectorial con las obras del proyecto.}
}
\value{
tibble con los datos preparados
}
\description{
Preparar datos
}
