% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp2kmz.R
\name{shp2kmz}
\alias{shp2kmz}
\title{Convertir Shapefile a KMZ con atributos extendidos}
\usage{
shp2kmz(shp, name = NULL, folder = NULL, dirsave = NULL)
}
\arguments{
\item{shp}{Ruta al archivo shapefile (.shp).}

\item{name}{Nombre de la columna para usar como etiqueta (Name). Si es NULL, usa la primera.}

\item{folder}{Columna opcional para agrupar en carpetas (Folder).}

\item{dirsave}{Directorio de salida. Por defecto, el mismo que el input.}
}
\value{
Genera un archivo .kmz y muestra un mensaje con la ruta.
}
\description{
Lee un archivo shapefile, lo proyecta a WGS84 y genera un archivo KMZ.
Mantiene los atributos originales en etiquetas `ExtendedData` permitiendo
su visualización en Google Earth. Permite además agrupar los elementos en carpetas.
}
