% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_guia_fragmentacion.R
\docType{data}
\name{tabla_guia_fragmentacion}
\alias{tabla_guia_fragmentacion}
\title{tabla_guia_fragmentacion}
\format{
## `tabla_guia_fragmentacion`
A tibble with 65 rows and 6 columns:
\describe{
  \item{N}{N° del parámetro}
  \item{Escala_espacial}{Escala de impacto del parámetro}
  \item{Parámetro}{Nombre del parámetro}
  \item{Sigla}{Sigla asociada a los parámetros a escala de Hábitat (parámetros de fragmentación)}
  \item{Categorías}{Categorías con las que se evalúan los parámetros}
  \item{Valoración}{Valoración de las categorías}
}
}
\source{
adaptado de "Guía para la solicitud de excepcionalidad del artículo 19 de la Ley 20283 sobre Recuperación del Bosque Nativo y Fomento Forestal"
}
\usage{
tabla_guia_fragmentacion
}
\description{
Tabla de los 19 parámetros para analisis de amenaza de la especie objeto del PAS 150.
}
\keyword{datasets}
