% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{download_files}
\alias{download_files}
\title{download_files}
\usage{
download_files(x, name_save, dir_save)
}
\arguments{
\item{x}{objecto o lista de objetos. admite objetos 'data.frame', 'sf' y 'wbWorkbook'.}

\item{name_save}{vector de caracteres con los nombres de los objetos.}

\item{dir_save}{directorio donde guardar el o los objetos.}
}
\value{
archivos 'xlsx', 'shp' o 'zip'.
}
\description{
download_files
}
