% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portadas.R
\name{portada}
\alias{portada}
\alias{wb_portada_PAS150}
\alias{portada_opts}
\title{Portadas de apendices}
\usage{
wb_portada_PAS150(
  wb,
  apendice = c("inventarios", "biodiversidad", "fragmentacion"),
  nom_ssubc = NULL,
  opts = portada_opts()
)

portada_opts(nom_proj = NULL, logo = NULL, plantilla = "default")
}
\arguments{
\item{wb}{Workbook.}

\item{apendice}{apendices correspondiente. `inventarios`, `biodiversidad` o `fragmentacion`.}

\item{nom_ssubc}{nombre de la subsubcuenca.}

\item{opts}{Opciones para personalizar una portada. usar [portada_opts()].}

\item{nom_proj}{Nombre del proyecto.}

\item{logo}{Ruta del archivo de tipo imagen con el logo del cliente.}

\item{plantilla}{Plantilla realizada para proyectos en particular, disponible `KIM753`, `MLP612` y `default`.}
}
\value{
Workbook con portada

lista con ajustes de portada
}
\description{
Portadas de apendices

Ajuste de portada
}
