% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{prepare_uso_veg}
\alias{prepare_uso_veg}
\title{Preparar y estandarizar datos de Uso y Vegetación}
\usage{
prepare_uso_veg(x)
}
\arguments{
\item{x}{Objeto sf con la cartografía de uso y vegetacion de la cuenca. Tambien util para darle formato a las capas de BNP en la cuenca y de intervenir y alterar.}
}
\value{
Objeto sf con columnas renombradas y superficie calculada.
}
\description{
Preparar y estandarizar datos de Uso y Vegetación
}
